+++ shared-core/src/androidMain/kotlin/dev/sdkforge/template/core/Platform.android.kt
@@ -1,0 +1,6 @@
+package dev.sdkforge.template.core
+
+actual val currentPlatform: Platform = object : Platform {
+    override val name: String get() = "Android"
+    override val version: String get() = android.os.Build.VERSION.SDK_INT.toString()
+}
