+++ shared-core/src/androidUnitTest/kotlin/dev/sdkforge/template/core/PlatformTest.android.kt
@@ -1,0 +1,14 @@
+@file:Suppress("ktlint:standard:filename")
+
+package dev.sdkforge.template.core
+
+import org.junit.Assert.assertTrue
+import org.junit.Test
+
+class AndroidPlatformTest {
+
+    @Test
+    fun testPlatformDefinition() {
+        assertTrue("Check Android is mentioned", currentPlatform.name.contains("Android", ignoreCase = true))
+    }
+}
