+++ shared-core/src/commonMain/kotlin/dev/sdkforge/template/core/Platform.kt
@@ -1,0 +1,8 @@
+package dev.sdkforge.template.core
+
+interface Platform {
+    val name: String
+    val version: String
+}
+
+expect val currentPlatform: Platform
