+++ shared-core/src/commonTest/kotlin/dev/sdkforge/template/core/PlatformTest.kt
@@ -1,0 +1,15 @@
+@file:Suppress("ktlint:standard:filename")
+
+package dev.sdkforge.template.core
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class CommonPlatformTest {
+
+    @Test
+    fun testPlatformDefinition() {
+        assertTrue(currentPlatform.name.isNotBlank(), "Platform name is present")
+        assertTrue(currentPlatform.version.isNotBlank(), "Platform version is present")
+    }
+}
