+++ shared-core/src/iosMain/kotlin/dev/sdkforge/template/core/Platform.ios.kt
@@ -1,0 +1,8 @@
+package dev.sdkforge.template.core
+
+import platform.UIKit.UIDevice
+
+actual val currentPlatform: Platform = object : Platform {
+    override val name: String get() = UIDevice.currentDevice.systemName
+    override val version: String get() = UIDevice.currentDevice.systemVersion
+}
