--- shared/src/androidMain/kotlin/dev/sdkforge/template/Platform.android.kt
@@ -1,6 +1,0 @@
-package dev.sdkforge.template
-
-actual val currentPlatform: Platform = object : Platform {
-    override val name: String get() = "Android"
-    override val version: String get() = android.os.Build.VERSION.SDK_INT.toString()
-}
