+++ shared/src/commonMain/kotlin/dev/sdkforge/template/Library.kt
@@ -1,0 +1,5 @@
+package dev.sdkforge.template
+
+data object Library {
+    const val VERSION = "0.0.1"
+}
