--- shared/src/commonMain/kotlin/dev/sdkforge/template/Platform.kt
@@ -1,8 +1,0 @@
-package dev.sdkforge.template
-
-interface Platform {
-    val name: String
-    val version: String
-}
-
-expect val currentPlatform: Platform
