--- shared/src/iosMain/kotlin/dev/sdkforge/template/Platform.ios.kt
@@ -1,8 +1,0 @@
-package dev.sdkforge.template
-
-import platform.UIKit.UIDevice
-
-actual val currentPlatform: Platform = object : Platform {
-    override val name: String get() = UIDevice.currentDevice.systemName
-    override val version: String get() = UIDevice.currentDevice.systemVersion
-}
